requirements.txt
qrcode
pillow
pyyaml
base32-crockford
